requirements.txt
#torch
pyyaml
#torchvision
Pillow
tqdm
